+++ src/Internal/utils.php
@@ -10,3 +10,8 @@ function safeGetByName(?array $collection): Closure
 {
     return fn (string $name) => is_array($collection) && isset($collection[$name]) && $collection[$name];
 }
+
+function flatMap(callable $fnc, array $arr, array ...$arrays): array
+{
+    return array_merge([], ...array_map($fnc, $arr, ...$arrays));
+}